Show and Index views for Student are being rendered
A list of courses displays for each student show view, along with grade percentage pts, letter grade, grade points for each courses and a GPA for the current term

Work with tests for controllers - the student#show test does not pass. The @grade instance variable produces issues. When it is commented out, the error produced is related to the ruby code in the view